+++ src/libclient/tools/toolcontroller.h
@@ -35,9 +35,9 @@ class TransformTool;
 class ToolController final : public QObject {
 	Q_PROPERTY(QCursor activeToolCursor READ activeToolCursor()
 				   NOTIFY toolCursorChanged)
-	Q_PROPERTY(uint16_t activeLayer READ activeLayer WRITE setActiveLayer NOTIFY
+	Q_PROPERTY(int activeLayer READ activeLayer WRITE setActiveLayer NOTIFY
 				   activeLayerChanged)
-	Q_PROPERTY(uint16_t activeAnnotation READ activeAnnotation WRITE
+	Q_PROPERTY(int activeAnnotation READ activeAnnotation WRITE
 				   setActiveAnnotation NOTIFY activeAnnotationChanged)
 	Q_PROPERTY(brushes::ActiveBrush activeBrush READ activeBrush WRITE
 				   setActiveBrush NOTIFY activeBrushChanged)
@@ -81,11 +81,11 @@ public:
 	QCursor activeToolCursor() const;
 	Capabilities activeToolCapabilities() const;
 
-	void setActiveLayer(uint16_t id);
-	uint16_t activeLayer() const { return m_activeLayer; }
+	void setActiveLayer(int id);
+	int activeLayer() const { return m_activeLayer; }
 
-	void setActiveAnnotation(uint16_t id);
-	uint16_t activeAnnotation() const { return m_activeAnnotation; }
+	void setActiveAnnotation(int id);
+	int activeAnnotation() const { return m_activeAnnotation; }
 
 	void setSelectedLayers(const QSet<int> &selectedLayers);
 	const QSet<int> &selectedLayers() const { return m_selectedLayers; }
@@ -227,7 +227,7 @@ signals:
 	void toolCapabilitiesChanged(unsigned int capabilities);
 	void toolCursorChanged(const QCursor &cursor);
 	void activeLayerChanged(int layerId);
-	void activeAnnotationChanged(uint16_t annotationId);
+	void activeAnnotationChanged(int annotationId);
 	void activeBrushChanged(const brushes::ActiveBrush &);
 	void transformToolStateChanged(int mode, int handle, bool dragging);
 	void modelChanged(canvas::CanvasModel *model);
@@ -277,8 +277,8 @@ private:
 
 	brushes::ActiveBrush m_activebrush;
 	Tool *m_activeTool;
-	uint16_t m_activeLayer;
-	uint16_t m_activeAnnotation;
+	int m_activeLayer;
+	int m_activeAnnotation;
 	QSet<int> m_selectedLayers;
 	QColor m_foregroundColor;
 	bool m_drawing;
